# openapi.yaml
openapi: 3.0.3
info:
  title: Appliance Shop API - OpenAPI 3.0
  description: |-
    This is a Home Appliances Store Server based on the OpenAPI 3.0 specification. API for device management in Home Appliance Stores.

    Some useful links:
    - [The Appliance Shop repository](https://github.com/minhcopilot/Appliance-Shop-API)
    - [The source API definition for the Appliance Shop](https://github.com/minhcopilot/Appliance-Shop-API/openopenapi.yaml)

  contact:
    email: phanminh190502@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:9000
    description: Local server
...
paths:
...
  /article/posts:
    get:
      tags:
        - post
      summary: Get all published posts
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
          description: Number of items per page
        - in: query
          name: authorId
          schema:
            type: integer
          description: Filter by author ID
        - in: query
          name: category
          schema:
            type: string
          description: Filter by category
        - in: query
          name: search
          schema:
            type: string
          description: Search term
        - in: query
          name: type
          schema:
            type: string
          description: Filter by post type
        - in: query
          name: sort
          schema:
            type: string
          description: Sort order
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  docs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
...
  /article/posts/all:
    get:
      security:
        - bearerAuth: []
      tags:
        - post
      summary: Get all posts (admin only)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
  /article/posts/all/search/query:
    get:
      security:
        - bearerAuth: []
      tags:
        - post
      summary: Search posts (admin only)
      parameters:
        - in: query
          name: query
          schema:
            type: string
          description: Search query
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
  /article/posts/{url}:
    get:
      tags:
        - post
      summary: Get post by URL
      parameters:
        - in: path
          name: url
          required: true
          schema:
            type: string
          description: Post URL
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
  /article/posts/all/{url}:
    get:
      security:
        - bearerAuth: []
      tags:
        - post
      summary: Get post by URL or ID (admin only)
      parameters:
        - in: path
          name: url
          required: true
          schema:
            type: string
          description: Post URL or ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
    delete:
      security:
        - bearerAuth: []
      tags:
        - post
      summary: Delete a post by URL (admin only)
      parameters:
        - in: path
          name: url
          required: true
          schema:
            type: string
          description: Post URL
      responses:
        '200':
          description: Post deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /article/posts/{url}/comments:
    get:
      tags:
        - post
      summary: Get comments for a post by URL
      parameters:
        - in: path
          name: url
          required: true
          schema:
            type: string
          description: Post URL
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
    post:
      tags:
        - post
      summary: Post a comment
      parameters:
        - in: path
          name: url
          required: true
          schema:
            type: string
          description: Post URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
            example:
              content: 'This is a comment.'
              postId: '60d0fe4f5311236168a109ca'
              status: 'pending'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
  /article/posts/all/{url}/comments:
    get:
      security:
        - bearerAuth: []
      tags:
        - post
      summary: Get comments for a post by URL or ID (admin only)
      parameters:
        - in: path
          name: url
          required: true
          schema:
            type: string
          description: Post URL or ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
  /article/posts/{url}/like:
    post:
      tags:
        - post
      summary: Like or unlike a post
      parameters:
        - in: path
          name: url
          required: true
          schema:
            type: string
          description: Post URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                like:
                  type: boolean
              example:
                like: true
      responses:
        '200':
          description: Like status
          content:
            application/json:
              schema:
                type: boolean
  /article/posts/all/:
    post:
      security:
        - bearerAuth: []
      tags:
        - post
      summary: Create a post (admin only)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
                url:
                  type: string
                status:
                  type: string
                  enum: [published, draft]
                category:
                  type: string
                file:
                  type: string
                  format: binary
              example:
                title: 'Sample Post'
                content: 'This is a sample post.'
                url: 'sample-post'
                status: 'published'
                category: 'sample-category'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
  /article/posts/all/{id}:
    patch:
      security:
        - bearerAuth: []
      tags:
        - post
      summary: Update a post by ID (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Post ID
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
                url:
                  type: string
                status:
                  type: string
                  enum: [published, draft]
                category:
                  type: string
                file:
                  type: string
                  format: binary
              example:
                title: 'Updated Sample Post'
                content: 'This is an updated sample post.'
                url: 'updated-sample-post'
                status: 'published'
                category: 'updated-sample-category'
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'

  #Chat
...
components:
  schemas:
...
    Comment:
      type: object
      properties:
        _id:
          type: string
          example: '60d0fe4f5311236168a109ca'
        content:
          type: string
          example: 'This is a comment'
        post:
          type: string
          example: '60d0fe4f5311236168a109cb'
        createdAt:
          type: string
          format: date-time
          example: '2021-06-21T17:32:28Z'
        updatedAt:
          type: string
          format: date-time
          example: '2021-06-21T17:32:28Z'
    Post:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [published, draft]
        authorId:
          type: number
        authorName:
          type: string
        category:
          type: string
        imageUrl:
          type: object
          properties:
            url:
              type: string
            publicId:
              type: string
            name:
              type: string
            size:
              type: number
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT